Move etl_stats test code
